third_party/blink/web_tests/external/wpt/css/cssom-view/mouseEvent-expected.txt

This is a testharness.js-based test.
[FAIL] MouseEvent's pageX and pageY attributes should be the sum of the scroll offset and clientX/clientY
  assert_equals: expected 5020 but got 20
Harness: the test ran to completion.
